daemon.py: application module
