+++ localslackirc/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+from datetime import datetime
+import pkgutil
+
+__path__ = pkgutil.extend_path(__path__, __name__)
+
+__title__ = 'localslickirc'
+__version__ = '2.0'
+__author__ = 'Salvo "LtWorf" Tomaselli'
+__copyright__ = 'Copyright(C) 2010-%s Salvo "LtWorf" Tomaselli' % datetime.today().year
