+++ localslackirc/__main__.py
@@ -1,0 +1,3 @@
+from .daemon import Daemon
+
+Daemon.run()
